bug(LottoResult): 수익률 타입 double로 변경

+++ src/main/java/lotto/LottoResult.java
@@ -2,5 +2,5 @@ package lotto;
 
 import java.util.List;
 
-public record LottoResult(List<MatchingCountResult> matchingCountResults, int rate) {
+public record LottoResult(List<MatchingCountResult> matchingCountResults, double rate) {
 }
